// src/enums/exception.enum.ts
enum Exception {
  INVALID_FIELD = 'Campo inválido.',
  NOT_FOUND = 'Nenhum resultado encontrado.',
  REQUIRED_FIELD = 'Campo obrigatório.',
  REQUIRED_AUTHENTICATION = 'Faça o login para prosseguir.',
  PERMISSION_DENIED = 'Ação não-autorizada.',
  EXISTING_EMAIL = 'Outra conta está utilizando esse e-mail.',
}

export default Exception;